1. Minus button when trivit is expanded - color ideas or white is fine?
2. DEV: simulate crash only available is you add a trivit with name: ThisIsADevDevice
To remove from settings, rename a trivit to: ThisIsNoDevDeviceNoMore

3. rename has less flickering because of resizing views
4.already a bit of less redrawing

+++ trivit WatchKit Extension/WKTableVIewRowController.m
@@ -25,7 +25,7 @@
     [self.completeCell setBackgroundColor:color];
 }
 - (void)setCounter:(NSInteger)count{
-    [self.itemCountButton setTitle:[NSString stringWithFormat:@"%ld",count ]];
+    [self.itemCountButton setTitle:[NSString stringWithFormat:@"%ld",(long)count ]];
 }
 
 @end
